+++ concordium-consensus/src/Concordium/MultiVersion.hs
@@ -49,6 +49,7 @@ import Concordium.Afgjort.Finalize
 import Concordium.Afgjort.Finalize.Types
 import Concordium.Birk.Bake
 import Concordium.GlobalState
+import qualified Concordium.GlobalState.AccountMap.LMDB as LMDBAccountMap
 import Concordium.GlobalState.Block
 import Concordium.GlobalState.BlockPointer (BlockPointer (..), BlockPointerData (..))
 import Concordium.GlobalState.Finalization
@@ -166,10 +167,18 @@ instance
 
 -- | Configuration for the global state that uses disk storage
 --  for both tree state and block state.
-newtype DiskStateConfig = DiskStateConfig
+data DiskStateConfig = DiskStateConfig
     { -- | Root directory for the global state.
-      stateBasePath :: FilePath
+      stateBasePath :: !FilePath,
+      -- | Account map.
+      globalAccountMap :: !LMDBAccountMap.DatabaseHandlers
     }
+
+-- | Create a 'DiskStateConfig', opening the shared account map database in the process.
+makeDiskStateConfig :: FilePath -> IO DiskStateConfig
+makeDiskStateConfig stateBasePath = do
+    globalAccountMap <- LMDBAccountMap.openDatabase (stateBasePath </> "accountmap")
+    return DiskStateConfig{..}
 
 -- | Configuration information for a multi-version runner.
 --  The type parameter defines the finalization configuration, and should be an instance of
@@ -188,15 +197,13 @@ globalStateConfig ::
     DiskStateConfig ->
     RuntimeParameters ->
     GenesisIndex ->
-    -- | Absolute height of the genesis block.
-    AbsoluteBlockHeight ->
     GlobalStateConfig
-globalStateConfig DiskStateConfig{..} rtp gi _ =
+globalStateConfig DiskStateConfig{..} rtp gi =
     ( GlobalStateConfig
         { dtdbRuntimeParameters = rtp,
           dtdbTreeStateDirectory = stateBasePath </> ("treestate-" ++ show gi),
           dtdbBlockStateFile = stateBasePath </> ("blockstate-" ++ show gi) <.> "dat",
-          dtdAccountMapDirectory = stateBasePath </> "accountmap"
+          gscAccountMap = globalAccountMap
         }
     )
 
@@ -211,7 +218,7 @@ globalStateConfigV1 DiskStateConfig{..} rtp gi =
         { gscRuntimeParameters = rtp,
           gscTreeStateDirectory = stateBasePath </> ("treestate-" ++ show gi),
           gscBlockStateFile = stateBasePath </> ("blockstate-" ++ show gi) <.> "dat",
-          gscAccountMapDirectory = stateBasePath </> "accountmap"
+          gscAccountMap = globalAccountMap
         }
     )
 
@@ -639,7 +646,6 @@ newGenesis (PVGenesisData (gd :: GenesisData pv)) genesisHeight = case consensus
                                         mvcStateConfig
                                         mvcRuntimeParameters
                                         vc0Index
-                                        genesisHeight
                                     )
                                     mvcFinalizationConfig
                                     UpdateHandler
@@ -747,7 +753,6 @@ checkForProtocolUpdateV0 = liftSkov body
                                         (mvcStateConfig mvConfiguration)
                                         (mvcRuntimeParameters mvConfiguration)
                                         vc0Index
-                                        vc0GenesisHeight
                                     )
                                     (mvcFinalizationConfig mvConfiguration)
                                     UpdateHandler
@@ -1193,7 +1198,6 @@ startupSkov genesis = do
                                     mvcStateConfig
                                     mvcRuntimeParameters
                                     genIndex
-                                    genHeight
                                 )
                                 mvcFinalizationConfig
                                 UpdateHandler
@@ -1457,7 +1461,10 @@ shutdownMultiVersionRunner MultiVersionRunner{..} = mask_ $ do
     -- Acquire the write lock. This prevents further updates, as they will block.
     takeMVar mvWriteLock
     versions <- readIORef mvVersions
+    -- Shut down the consensus databases.
     runLoggerT (forM_ versions evcShutdown) mvLog
+    -- Shut down the global account map.
+    LMDBAccountMap.closeDatabase (globalAccountMap (mvcStateConfig mvConfiguration))
 
 -- | Lift a version-0 consensus skov action to the 'MVR' monad, running it on a
 --  particular 'VersionedConfigurationV0'. Note that this does not